Initial commit for ESP8266 and ESP32
This library has been tested with ESP8266 and ESP32. Some filenames that had
8266 in the name have had the 8266 removed to reflect that.

Compilation was done with g++ -Wall option which explains why many char *
have been changed to const char *. Also -Wall showed us some local variable
errors in AWSClient4.cpp and we added some class members to fix that.

I added a content-length header to canonical headers only because I was
trying to follow AWS instructions literally. What really matters is that the
signedHeaders set in AmazonIOTClient corresponds to those used.

A couple of places I conditionally compile against ESP32 and/or ESP8266.
Mainly that is to select which header files to include, but there is a
functional thing in EspHttpClient::send that does not work for ESP32 as
written so I left it out.

I had to change the receive logic of that function to receive the entire
response and close the persistent connection in a simple manner. Likewise, I
modified updateCurlTime to avoid using readString() from the underlying
stream.